lo intetamos

+++ src/main/java/com/asj/listed/repositories/Repositories.java
@@ -4,8 +4,8 @@ import java.util.List;
 import java.util.Optional;
 public interface Repositories<T> {
     List<T> listarTodos();
-    Optional<T> buscarPorId(int id);
+    Optional<T> buscarPorId(long id);
     T crear(T t);
-    T actualizar (int id,T t);
-    Optional<T> eliminar (int id);
+    T actualizar (long id,T t);
+    Optional<T> eliminar (long id);
 }
